+++ packages/react-i18n/CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ---
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+- Added support for pluralization to `getTranslationTree` ([#988](https://github.com/Shopify/quilt/pull/988))
 
 ## [1.8.3] - 2019-09-03